moves extensions code to UtilityLib
