+++ VaxineApp/VaxineApp/ViewModels/Home/Area/Doctor/DoctorViewModel.cs
@@ -5,7 +5,7 @@ using System.Collections.ObjectModel;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
-using VaxineApp.Helpers;
+using UtilityLib.Extensions;
 using VaxineApp.Models;
 using VaxineApp.MVVMHelper;
 using VaxineApp.StaticData;
